requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Machine Learning & NLP
sentence-transformers==2.7.0
torch==2.4.1
transformers==4.44.2
tokenizers==0.19.1

# Vector Database
pinecone-client==3.0.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# CLI (for backward compatibility)
click==8.1.7

# Additional utilities
numpy==2.1.3

# Database (MongoDB)
pymongo==4.8.0

# Auth
Flask-JWT-Extended==4.6.0

# Build tooling (ensure wheels can be built when needed)
setuptools>=68.0.0
wheel>=0.41.0
